Use default if there is no template for the role.

+++ plugins/shares/netlogon/open/oss-create-logon-script.sh
@@ -12,7 +12,11 @@ mkdir -p /var/lib/samba/sysvol/$R/scripts
 setfacl -m g:users:rx /var/lib/samba/sysvol/
 setfacl -m g:users:rx /var/lib/samba/sysvol/$R/
 setfacl -m g:users:rx /var/lib/samba/sysvol/$R/scripts/
-cp /usr/share/oss/templates/login-${role}.bat /var/lib/samba/sysvol/$R/scripts/${U}.bat
+if [ -e /usr/share/oss/templates/login-${role}.bat ]; then
+	cp /usr/share/oss/templates/login-${role}.bat /var/lib/samba/sysvol/$R/scripts/${U}.bat
+else
+	cp /usr/share/oss/templates/login-default.bat /var/lib/samba/sysvol/$R/scripts/${U}.bat
+fi
 chown ${U} /var/lib/samba/sysvol/$R/scripts/${U}.bat
 setfacl -m m::rwx /var/lib/samba/sysvol/$R/scripts/${U}.bat
